moved config into scripts
