+++ api_tests/tests/conftest.py
@@ -4,7 +4,6 @@ from api_test_utils.oauth_helper import OauthHelper
 from api_test_utils.apigee_api_apps import ApigeeApiDeveloperApps
 from api_test_utils.apigee_api_products import ApigeeApiProducts
 from api_tests.scripts.generic_request import GenericRequest
-from api_tests.config_files.config import SERVICE_NAME
 
 
 @pytest.fixture()
