+++ api_tests/tests/pds/test_asid_parameter.py
@@ -1,5 +1,4 @@
 import pytest
-from api_tests.config_files import config
 
 
 @pytest.mark.skip(reason="Design Discussion Required")
